Refactor export card
