+++ src/apps/companies/apps/company-overview/overview-table-cards/index.jsx
@@ -3,4 +3,4 @@ export { default as ActivityCard } from './ActivityCard'
 export { default as BusinessDetailsCard } from './BusinessDetailsCard'
 export { default as ActiveInvestmentProjectsCard } from './ActiveInvestmentProjectsCard'
 export { default as InvestmentStatusCard } from './InvestmentStatusCard'
-export { default as ExportStatusCard } from './ExportStatus'
+export { default as ExportStatusCard } from './ExportStatusCard'
